Seperated wordOfTheDay and randomWord tests
