+++ lib/tests/wordOfTheDay-test.js
@@ -2,18 +2,7 @@
 
 import { assertExists, assertEquals } from 'assert'
 
-import { randomWord, wordOfTheDay } from '../mod.js'
-
-Deno.test({
-	name: 'words/randomWord',
-	permissions: { read: true, env: true, net: true },
-	async fn() {
-		const response = await randomWord()
-		assertExists(response)
-		assertExists(response.word)
-		console.log(response.word)
-	}
-})
+import { wordOfTheDay } from '../mod.js'
 
 Deno.test({
 	name: 'words/wordOfTheDay: today',
